+++ src/app/common/business-models/lease-contract.business-model.ts
@@ -5,6 +5,8 @@ import {PropertyBusinessModel} from "./property.business-model";
 export interface LeaseContractBusinessModel {
   id: string;
   rentAmount: number;
+  rentCharges: number;
+  dateContractSignature: Date;
   owner: OwnerBusinessModel;
   tenant: TenantBusinessModel;
   property: PropertyBusinessModel;