+++ src/app/features/lease-contract/lease-contract.module.ts
@@ -1,5 +1,5 @@
 import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
+import {CommonModule, DatePipe} from '@angular/common';
 import {CreateLeaseContractComponent} from './create-lease-contract/create-lease-contract.component';
 import {UpdateLeaseContractComponent} from './update-lease-contract/update-lease-contract.component';
 import {LeaseContractDetailComponent} from "./lease-contract-detail/lease-contract-detail.component";
@@ -12,6 +12,8 @@ import {MatSelectModule} from "@angular/material/select";
 import {LeaseContractFormComponent} from './lease-contract-form/lease-contract-form.component';
 import {ReactiveFormsModule} from "@angular/forms";
 import {ComponentsModule} from "../../common/components/components.module";
+import {MatDatepickerModule} from "@angular/material/datepicker";
+import {MatIconModule} from "@angular/material/icon";
 
 
 @NgModule({
@@ -30,7 +32,12 @@ import {ComponentsModule} from "../../common/components/components.module";
     MatInputModule,
     MatSelectModule,
     ReactiveFormsModule,
-    ComponentsModule
+    ComponentsModule,
+    MatDatepickerModule,
+    MatIconModule
+  ],
+  providers: [
+    DatePipe
   ]
 })
 export class LeaseContractModule {