+++ CHANGELOG.md
@@ -14,6 +14,12 @@ Nomenclature :
 
 ### Unreleased
 
+#### Added
+
+- Updated db.json with data to complete the simulated api with lease-contract data
+- Added GET /lease-contracts/:id and display it in a card
+- Added POST /lease-contracts (form)
+
 ### [0.1.2] - 2023-03-15
 
 #### Added